// libs/features/benefits/src/lib/shared/enum/work-flow-type.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export enum WorkFlowType {
  REQUEST_UNEMPLOYMENT = 'REQUEST_UNEMPLOYMENT',
  REJECT_SANED = 'Reject Saned',
  REQUEST_INSPECTION = 'Request Inspection',
  REQUEST_WOMAN_LUMPSUM = 'REQUEST_WOMAN_LUMPSUM',
  REQUEST_RETIREMENT_LUMPSUM = 'REQUEST_RETIREMENT_LUMPSUM',
  REQUEST_RETIREMENT_PENSION = 'REQUEST_RETIREMENT_PENSION',
  REQUEST_EARLY_RETIRMENT_PENSION = 'REQUEST_EARLY_RETIREMENT_PENSION',
  SANED = 'Request Unemployment',
  PROACTIVE_SANED = 'Proactive Saned Request',
  REQUEST_NON_OCCUPATIONAL_BENEFIT = 'REQUEST_NON_OCCUPATIONAL_BENEFIT',
  REQUEST_NON_OCC_LUMPSUM = 'REQUEST_NON_OCC_LUMPSUM',
  REQUEST_NON_OCC_PENSION = 'REQUEST_NON_OCC_PENSION',
  REQUEST_JAILED_CONTRIBUTOR_PENSION = 'REQUEST_JAILED_CONTRIBUTOR_PENSION',
  REQUEST_JAILED_CONTRIBUTOR_LUMPSUM = 'REQUEST_JAILED_CONTRIBUTOR_LUMPSUM',
  REQUEST_HAZARDOUS_PENSION = 'REQUEST_HAZARDOUS_PENSION',
  REQUEST_HAZARDOUS_LUMPSUM = 'REQUEST_HAZARDOUS_LUMPSUM',
  REQUEST_HEIR_PENSION = 'REQUEST_HEIR_PENSION',
  REQUEST_HEIR_LUMPSUM = 'REQUEST_HEIR_LUMPSUM',
  REQUEST_OCC_PENSION = 'REQUEST_OCCUPATIONAL_PENSION',
  REQUEST_OCC_LUMPSUM = 'REQUEST_OCCUPATIONAL_LUMPSUM',
  REQUEST_MODIFY_BENEFIT = 'MODIFY_DEPENDENT',
  REQUEST_ADD_BENEFIT = 'REQUEST_ADD_RETIREMENT_PENSION',
  REQUEST_MODIFY_HEIR = 'MODIFY_HEIR',
  REQUEST_ADD_HEIR = 'ADD_HEIR',
  REQUEST_RETURN_LUMPSUM = 'RET_LUMPSUM_BEN',
  REGISTER_HEIR = 'REGISTER_HEIR',
  REQUEST_FUNERAL_GRANT = 'REQUEST_FUNERAL_GRANT'
}