// libs/features/benefits/src/lib/shared/enum/uipayload-key-enum.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export enum UIPayloadKeyEnum {
  SIN = 'socialInsuranceNo',
  ID = 'id',
  PERSON_ID = 'PersonId',
  REFERENCE_NO = 'referenceNo',
  RESOURCE = 'resource',
  SUB_RESOURCE ='subResource',
  CHANNEL = 'sourcechannel',
  EVENT_NAME = 'eventName',
  REPAY_ID = 'repayId'
}